requirements.txt
Flask>=2.0
python-dotenv>=0.19
requests>=2.25
PyJWT>=2.0
supabase>=1.0 